+++ diesel/src/sqlite/types/date_and_time/chrono.rs
@@ -2,45 +2,18 @@ extern crate chrono;
 
 use std::error::Error;
 use std::io::Write;
-use self::chrono::{Datelike, NaiveDate, NaiveDateTime, NaiveTime};
+use self::chrono::{NaiveDate, NaiveDateTime, NaiveTime};
 
 use sqlite::Sqlite;
 use sqlite::connection::SqliteValue;
 use types::{Date, FromSql, IsNull, Text, Time, Timestamp, ToSql, ToSqlOutput};
 
-const SQLITE_TIME_FMT: &'static str = "%0H:%0M:%0S%.6f";
-
-fn parse_sqlite_date(date: &str) -> Result<NaiveDate, Box<Error + Send + Sync>> {
-    let negative_year = date.starts_with('-');
-
-    let ymd: Vec<u32> = date.split('-').flat_map(str::parse).collect();
-    if ymd.len() != 3 {
-        return Err(
-            "Cannot parse date, too many parts. The date should be in the form: YYYY-MM-DD".into(),
-        );
-    }
-
-    if negative_year {
-        return Ok(NaiveDate::from_ymd(-(ymd[0] as i32), ymd[1], ymd[2]));
-    }
-
-    NaiveDate::from_ymd_opt(ymd[0] as i32, ymd[1], ymd[2])
-        .ok_or_else(|| format!("Cannot parse this date: {:?}", ymd).into())
-}
-
-fn dump_sqlite_date<T: Datelike>(date: &T) -> String {
-    format!(
-        "{:0>4}-{:0>2}-{:0>2}",
-        date.year(),
-        date.month(),
-        date.day()
-    )
-}
+const SQLITE_DATE_FORMAT: &'static str = "%F";
 
 impl FromSql<Date, Sqlite> for NaiveDate {
     fn from_sql(value: Option<&SqliteValue>) -> Result<Self, Box<Error + Send + Sync>> {
         let text = not_none!(value).read_text();
-        parse_sqlite_date(text)
+        Self::parse_from_str(text, SQLITE_DATE_FORMAT).map_err(Into::into)
     }
 }
 
@@ -49,7 +22,7 @@ impl ToSql<Date, Sqlite> for NaiveDate {
         &self,
         out: &mut ToSqlOutput<W, Sqlite>,
     ) -> Result<IsNull, Box<Error + Send + Sync>> {
-        let s = dump_sqlite_date(self);
+        let s = self.format(SQLITE_DATE_FORMAT).to_string();
         ToSql::<Text, Sqlite>::to_sql(&s, out)
     }
 }
@@ -57,8 +30,24 @@ impl ToSql<Date, Sqlite> for NaiveDate {
 impl FromSql<Time, Sqlite> for NaiveTime {
     fn from_sql(value: Option<&SqliteValue>) -> Result<Self, Box<Error + Send + Sync>> {
         let text = not_none!(value).read_text();
-        let time = NaiveTime::parse_from_str(text, SQLITE_TIME_FMT)?;
-        Ok(time)
+        let valid_time_formats = &[
+            // Most likely
+            "%T%.f",
+            // All other valid formats in order of documentation
+            "%R",
+            "%RZ",
+            "%T%.fZ",
+            "%R%:z",
+            "%T%.f%:z",
+        ];
+
+        for format in valid_time_formats {
+            if let Ok(time) = Self::parse_from_str(text, format) {
+                return Ok(time);
+            }
+        }
+
+        Err(format!("Invalid time {}", text).into())
     }
 }
 
@@ -67,7 +56,7 @@ impl ToSql<Time, Sqlite> for NaiveTime {
         &self,
         out: &mut ToSqlOutput<W, Sqlite>,
     ) -> Result<IsNull, Box<Error + Send + Sync>> {
-        let s = format!("{}", self.format(SQLITE_TIME_FMT));
+        let s = self.format("%T%.f").to_string();
         ToSql::<Text, Sqlite>::to_sql(&s, out)
     }
 }
@@ -75,16 +64,42 @@ impl ToSql<Time, Sqlite> for NaiveTime {
 impl FromSql<Timestamp, Sqlite> for NaiveDateTime {
     fn from_sql(value: Option<&SqliteValue>) -> Result<Self, Box<Error + Send + Sync>> {
         let text = not_none!(value).read_text();
-        let dt: Vec<&str> = text.split(' ').collect();
 
-        if dt.len() != 2 {
-            return Err("Can't parse Datetime. Too many parts. It should be in the form: YYYY-MM-DD HH:MM:SS".into());
+        let sqlite_datetime_formats = &[
+            // Most likely format
+            "%F %T%.f",
+            // Other formats in order of appearance in docs
+            "%F %R",
+            "%F %RZ",
+            "%F %R%:z",
+            "%F %T%.fZ",
+            "%F %T%.f%:z",
+            "%FT%R",
+            "%FT%RZ",
+            "%FT%R%:z",
+            "%FT%T%.f",
+            "%FT%T%.fZ",
+            "%FT%T%.f%:z",
+        ];
+
+        for format in sqlite_datetime_formats {
+            if let Ok(dt) = Self::parse_from_str(text, format) {
+                return Ok(dt);
+            }
         }
 
-        let time = NaiveTime::parse_from_str(dt[1], SQLITE_TIME_FMT)?;
-        let datetime = parse_sqlite_date(dt[0])?.and_time(time);
+        if let Ok(julian_days) = text.parse::<f64>() {
+            let epoch_in_julian_days = 2440587.5;
+            let seconds_in_day = 86400.0;
+            let timestamp = (julian_days - epoch_in_julian_days) * seconds_in_day;
+            let seconds = timestamp as i64;
+            let nanos = (timestamp.fract() * 1E9) as u32;
+            if let Some(timestamp) = Self::from_timestamp_opt(seconds, nanos) {
+                return Ok(timestamp);
+            }
+        }
 
-        Ok(datetime)
+        Err(format!("Invalid datetime {}", text).into())
     }
 }
 
@@ -93,9 +108,8 @@ impl ToSql<Timestamp, Sqlite> for NaiveDateTime {
         &self,
         out: &mut ToSqlOutput<W, Sqlite>,
     ) -> Result<IsNull, Box<Error + Send + Sync>> {
-        ToSql::<Date, Sqlite>::to_sql(&self.date(), out)?;
-        write!(out, " ")?;
-        ToSql::<Time, Sqlite>::to_sql(&self.time(), out)
+        let s = self.format("%F %T%.f").to_string();
+        ToSql::<Text, Sqlite>::to_sql(&s, out)
     }
 }
 
@@ -106,13 +120,16 @@ mod tests {
     extern crate dotenv;
 
     use self::chrono::{Duration, NaiveDate, NaiveDateTime, NaiveTime, Timelike, Utc};
-    use self::chrono::naive::MAX_DATE;
     use self::dotenv::dotenv;
 
     use select;
     use dsl::{now, sql};
     use prelude::*;
-    use types::{Date, Time, Timestamp};
+    use types::{Date, Text, Time, Timestamp};
+
+    sql_function!(datetime, datetime_t, (x: Text) -> Timestamp);
+    sql_function!(time, time_t, (x: Text) -> Time);
+    sql_function!(date, date_t, (x: Text) -> Date);
 
     fn connection() -> SqliteConnection {
         dotenv().ok();
@@ -127,18 +144,63 @@ mod tests {
     fn unix_epoch_encodes_correctly() {
         let connection = connection();
         let time = NaiveDate::from_ymd(1970, 1, 1).and_hms(0, 0, 0);
-        let query = select(sql::<Timestamp>("'1970-01-01 00:00:00.000000'").eq(time));
-        assert!(query.get_result::<bool>(&connection).unwrap());
+        let query = select(datetime("1970-01-01 00:00:00.000000").eq(time));
+        assert_eq!(Ok(true), query.get_result(&connection));
     }
 
-
     #[test]
-    fn unix_epoch_decodes_correctly() {
+    fn unix_epoch_decodes_correctly_in_all_possible_formats() {
         let connection = connection();
         let time = NaiveDate::from_ymd(1970, 1, 1).and_hms(0, 0, 0);
-        let epoch_from_sql =
-            select(sql::<Timestamp>("'1970-1-1 00:00:00'")).get_result(&connection);
-        assert_eq!(Ok(time), epoch_from_sql);
+        let valid_epoch_formats = vec![
+            "1970-01-01 00:00",
+            "1970-01-01 00:00:00",
+            "1970-01-01 00:00:00.000",
+            "1970-01-01 00:00:00.000000",
+            "1970-01-01T00:00",
+            "1970-01-01T00:00:00",
+            "1970-01-01T00:00:00.000",
+            "1970-01-01T00:00:00.000000",
+            "1970-01-01 00:00Z",
+            "1970-01-01 00:00:00Z",
+            "1970-01-01 00:00:00.000Z",
+            "1970-01-01 00:00:00.000000Z",
+            "1970-01-01T00:00Z",
+            "1970-01-01T00:00:00Z",
+            "1970-01-01T00:00:00.000Z",
+            "1970-01-01T00:00:00.000000Z",
+            "1970-01-01 00:00+00:00",
+            "1970-01-01 00:00:00+00:00",
+            "1970-01-01 00:00:00.000+00:00",
+            "1970-01-01 00:00:00.000000+00:00",
+            "1970-01-01T00:00+00:00",
+            "1970-01-01T00:00:00+00:00",
+            "1970-01-01T00:00:00.000+00:00",
+            "1970-01-01T00:00:00.000000+00:00",
+            "1970-01-01 00:00+01:00",
+            "1970-01-01 00:00:00+01:00",
+            "1970-01-01 00:00:00.000+01:00",
+            "1970-01-01 00:00:00.000000+01:00",
+            "1970-01-01T00:00+01:00",
+            "1970-01-01T00:00:00+01:00",
+            "1970-01-01T00:00:00.000+01:00",
+            "1970-01-01T00:00:00.000000+01:00",
+            "1970-01-01T00:00-01:00",
+            "1970-01-01T00:00:00-01:00",
+            "1970-01-01T00:00:00.000-01:00",
+            "1970-01-01T00:00:00.000000-01:00",
+            "1970-01-01T00:00-01:00",
+            "1970-01-01T00:00:00-01:00",
+            "1970-01-01T00:00:00.000-01:00",
+            "1970-01-01T00:00:00.000000-01:00",
+            "2440587.5",
+        ];
+
+        for s in valid_epoch_formats {
+            let epoch_from_sql =
+                select(sql::<Timestamp>(&format!("'{}'", s))).get_result(&connection);
+            assert_eq!(Ok(time), epoch_from_sql, "format {} failed", s);
+        }
     }
 
     #[test]
@@ -146,11 +208,11 @@ mod tests {
         let connection = connection();
         let time = Utc::now().naive_utc() + Duration::seconds(60);
         let query = select(now.lt(time));
-        assert!(query.get_result::<bool>(&connection).unwrap());
+        assert_eq!(Ok(true), query.get_result(&connection));
 
         let time = Utc::now().naive_utc() - Duration::seconds(600);
         let query = select(now.gt(time));
-        assert!(query.get_result::<bool>(&connection).unwrap());
+        assert_eq!(Ok(true), query.get_result(&connection));
     }
 
     #[test]
@@ -158,11 +220,11 @@ mod tests {
         let connection = connection();
 
         let midnight = NaiveTime::from_hms(0, 0, 0);
-        let query = select(sql::<Time>("'00:00:00.000000'").eq(midnight));
+        let query = select(time("00:00:00.000000").eq(midnight));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
         let noon = NaiveTime::from_hms(12, 0, 0);
-        let query = select(sql::<Time>("'12:00:00.000000'").eq(noon));
+        let query = select(time("12:00:00.000000").eq(noon));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
         let roughly_half_past_eleven = NaiveTime::from_hms_micro(23, 37, 4, 2200);
@@ -174,8 +236,37 @@ mod tests {
     fn times_of_day_decode_correctly() {
         let connection = connection();
         let midnight = NaiveTime::from_hms(0, 0, 0);
-        let query = select(sql::<Time>("'00:00:00'"));
-        assert_eq!(Ok(midnight), query.get_result::<NaiveTime>(&connection));
+        let valid_midnight_formats = &[
+            "00:00",
+            "00:00:00",
+            "00:00:00.000",
+            "00:00:00.000000",
+            "00:00Z",
+            "00:00:00Z",
+            "00:00:00.000Z",
+            "00:00:00.000000Z",
+            "00:00+00:00",
+            "00:00:00+00:00",
+            "00:00:00.000+00:00",
+            "00:00:00.000000+00:00",
+            "00:00+01:00",
+            "00:00:00+01:00",
+            "00:00:00.000+01:00",
+            "00:00:00.000000+01:00",
+            "00:00-01:00",
+            "00:00:00-01:00",
+            "00:00:00.000-01:00",
+            "00:00:00.000000-01:00",
+        ];
+        for format in valid_midnight_formats {
+            let query = select(sql::<Time>(&format!("'{}'", format)));
+            assert_eq!(
+                Ok(midnight),
+                query.get_result::<NaiveTime>(&connection),
+                "format {} failed",
+                format
+            );
+        }
 
         let noon = NaiveTime::from_hms(12, 0, 0);
         let query = select(sql::<Time>("'12:00:00'"));
@@ -193,23 +284,19 @@ mod tests {
     fn dates_encode_correctly() {
         let connection = connection();
         let january_first_2000 = NaiveDate::from_ymd(2000, 1, 1);
-        let query = select(sql::<Date>("'2000-01-01'").eq(january_first_2000));
+        let query = select(date("2000-01-01").eq(january_first_2000));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
-        let distant_past = NaiveDate::from_ymd(-398, 4, 11);
-        let query = select(sql::<Date>("'-398-04-11'").eq(distant_past));
-        assert!(query.get_result::<bool>(&connection).unwrap());
-
-        let max_date = MAX_DATE;
-        let query = select(sql::<Date>("'262143-12-31'").eq(max_date));
+        let distant_past = NaiveDate::from_ymd(0, 4, 11);
+        let query = select(date("0000-04-11").eq(distant_past));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
         let january_first_2018 = NaiveDate::from_ymd(2018, 1, 1);
-        let query = select(sql::<Date>("'2018-01-01'").eq(january_first_2018));
+        let query = select(date("2018-01-01").eq(january_first_2018));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
-        let distant_future = NaiveDate::from_ymd(72_400, 1, 8);
-        let query = select(sql::<Date>("'72400-01-08'").eq(distant_future));
+        let distant_future = NaiveDate::from_ymd(9999, 1, 8);
+        let query = select(date("9999-01-08").eq(distant_future));
         assert!(query.get_result::<bool>(&connection).unwrap());
     }
 
@@ -217,29 +304,25 @@ mod tests {
     fn dates_decode_correctly() {
         let connection = connection();
         let january_first_2000 = NaiveDate::from_ymd(2000, 1, 1);
-        let query = select(sql::<Date>("'2000-1-1'"));
+        let query = select(date("2000-01-01"));
         assert_eq!(
             Ok(january_first_2000),
             query.get_result::<NaiveDate>(&connection)
         );
 
-        let distant_past = NaiveDate::from_ymd(-399, 4, 11);
-        let query = select(sql::<Date>("'-399-04-11'"));
+        let distant_past = NaiveDate::from_ymd(0, 4, 11);
+        let query = select(date("0000-04-11"));
         assert_eq!(Ok(distant_past), query.get_result::<NaiveDate>(&connection));
 
-        let max_date = MAX_DATE;
-        let query = select(sql::<Date>("'262143-12-31'"));
-        assert_eq!(Ok(max_date), query.get_result::<NaiveDate>(&connection));
-
         let january_first_2018 = NaiveDate::from_ymd(2018, 1, 1);
-        let query = select(sql::<Date>("'2018-01-01'"));
+        let query = select(date("2018-01-01"));
         assert_eq!(
             Ok(january_first_2018),
             query.get_result::<NaiveDate>(&connection)
         );
 
-        let distant_future = NaiveDate::from_ymd(72_400, 1, 8);
-        let query = select(sql::<Date>("'72400-01-08'"));
+        let distant_future = NaiveDate::from_ymd(9999, 1, 8);
+        let query = select(date("9999-01-08"));
         assert_eq!(
             Ok(distant_future),
             query.get_result::<NaiveDate>(&connection)
@@ -250,31 +333,31 @@ mod tests {
     fn datetimes_decode_correctly() {
         let connection = connection();
         let january_first_2000 = NaiveDate::from_ymd(2000, 1, 1).and_hms(1, 1, 1);
-        let query = select(sql::<Timestamp>("'2000-1-1 01:01:01.000000'"));
+        let query = select(datetime("2000-01-01 01:01:01.000000"));
         assert_eq!(
             Ok(january_first_2000),
             query.get_result::<NaiveDateTime>(&connection)
         );
 
-        let distant_past = NaiveDate::from_ymd(-399, 4, 11).and_hms(2, 2, 2);
-        let query = select(sql::<Timestamp>("'-399-04-11 02:02:02.000000'"));
+        let distant_past = NaiveDate::from_ymd(0, 4, 11).and_hms(2, 2, 2);
+        let query = select(datetime("0000-04-11 02:02:02.000000"));
         assert_eq!(
             Ok(distant_past),
             query.get_result::<NaiveDateTime>(&connection)
         );
 
         let january_first_2018 = NaiveDate::from_ymd(2018, 1, 1);
-        let query = select(sql::<Date>("'2018-01-01'"));
+        let query = select(date("2018-01-01"));
         assert_eq!(
             Ok(january_first_2018),
             query.get_result::<NaiveDate>(&connection)
         );
 
-        let distant_future = NaiveDate::from_ymd(72_400, 1, 8)
+        let distant_future = NaiveDate::from_ymd(9999, 1, 8)
             .and_hms(23, 59, 59)
             .with_nanosecond(100_000)
             .unwrap();
-        let query = select(sql::<Timestamp>("'72400-01-08 23:59:59.000100'"));
+        let query = select(sql::<Timestamp>("'9999-01-08 23:59:59.000100'"));
         assert_eq!(
             Ok(distant_future),
             query.get_result::<NaiveDateTime>(&connection)
@@ -285,11 +368,11 @@ mod tests {
     fn datetimes_encode_correctly() {
         let connection = connection();
         let january_first_2000 = NaiveDate::from_ymd(2000, 1, 1).and_hms(0, 0, 0);
-        let query = select(sql::<Timestamp>("'2000-01-01 00:00:00.000000'").eq(january_first_2000));
+        let query = select(datetime("2000-01-01 00:00:00.000000").eq(january_first_2000));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
-        let distant_past = NaiveDate::from_ymd(-398, 4, 11).and_hms(20, 00, 20);
-        let query = select(sql::<Timestamp>("'-398-04-11 20:00:20.000000'").eq(distant_past));
+        let distant_past = NaiveDate::from_ymd(0, 4, 11).and_hms(20, 00, 20);
+        let query = select(datetime("0000-04-11 20:00:20.000000").eq(distant_past));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
         let january_first_2018 = NaiveDate::from_ymd(2018, 1, 1)
@@ -299,8 +382,8 @@ mod tests {
         let query = select(sql::<Timestamp>("'2018-01-01 12:00:00.000500'").eq(january_first_2018));
         assert!(query.get_result::<bool>(&connection).unwrap());
 
-        let distant_future = NaiveDate::from_ymd(72_400, 1, 8).and_hms(0, 0, 0);
-        let query = select(sql::<Timestamp>("'72400-01-08 00:00:00.000000'").eq(distant_future));
+        let distant_future = NaiveDate::from_ymd(9999, 1, 8).and_hms(0, 0, 0);
+        let query = select(datetime("9999-01-08 00:00:00.000000").eq(distant_future));
         assert!(query.get_result::<bool>(&connection).unwrap());
     }
 }